[grocery-tracker-poc] Adjust input box fill.
Expect visible to add delay between page reload and input box fill in
the hopes that it will make this less flaky. Sometimes, the fill fails
and the zipcode box shows an error about an invalid zip code. I think
this is because the fill happens before the javascript is done
propagating so the box input gets overwritten and messes everything up.