[grocery-tracker-poc] Move some shit around again for no reason
This was trying to put things in a convenience way for the Dockerfile.
Not entirely sure this is good, becuase later I'd want to add more
Dockerfiles possibly for other containers (like sql db or cronjob or
data volumes) and then finally add a compose file in here too, but we'll
deal with that later.