[grocery-tracker-poc] switch to patchright
Patchright is a fork of playwright that disables the runtime and console
APIs in chromium and then makes it so that each page executes in its own
isolated context.

This does not work with firefox (so I'm stuck only using this inside
docker). Seems to help with some of the diagnostic tests and cuts down
on entropy slightly.

The debug window on the pause menu is broken (this is a known issue in
the github page).