[grocery-tracker-poc] Add little etsl experiment snippet
If we choose a chromium user agent while using firefox, one of the
fingerprint anomalies is that the javascript eval() function will be a
different length when converted to a string.

We can monkey patch this in firefox to make it the same length as it
would be in chromium with the added code in the comments.

Saving this for future reference, but unlikely that we will actually
need it...