[nodriver-test] get nodriver working inside docker.
This actually also works on the host machine, but it uses your globally
installed instance of chrome (and nodriver only works with chrome).

Would have liked to mount the user data directory onto the host file
system, but it always makes the directory with root as the owner and
there is no way to change this. This can be solved by making a docker
compose file, but I didn't feel like it.