[botbrowser-test] Add playwright code.
Try to browser to diagnostics tests with this. Everything is set up, but
it looks like the demo profile does not let you create or browse to any
webpages. (it shuts down and then the python script fails because the
page object is closed)

Looks like you need a premium botbrowser profile, which is only
available to researchers and academics after they apply. Looking to see
if there is a way to get one of these now...